+++ src/partials/home-page/filters.html
@@ -40,7 +40,7 @@
       <!-- ! -->
       <div class="filter_field">
         <div class="select-param">
-          <p class="param-name search-input">A to Z</p>
+          <p class="param-name">A to Z</p>
         </div>
         <div class="filter_icon_position">
           <svg width="20" height="20" class="icon-filter">
